Moved service specific code out of shared codebase
