+++ cmd/web/main.go
@@ -7,13 +7,13 @@ import (
 	"time"
 
 	"github.com/fdemchenko/exchanger/cmd/web/internal/messaging"
+	"github.com/fdemchenko/exchanger/cmd/web/internal/repositories"
+	"github.com/fdemchenko/exchanger/cmd/web/internal/services"
+	"github.com/fdemchenko/exchanger/cmd/web/internal/services/rate"
 	"github.com/fdemchenko/exchanger/internal/communication/customers"
 	"github.com/fdemchenko/exchanger/internal/communication/mailer"
 	"github.com/fdemchenko/exchanger/internal/communication/rabbitmq"
 	"github.com/fdemchenko/exchanger/internal/database"
-	"github.com/fdemchenko/exchanger/internal/repositories"
-	"github.com/fdemchenko/exchanger/internal/services"
-	"github.com/fdemchenko/exchanger/internal/services/rate"
 	"github.com/fdemchenko/exchanger/migrations"
 	_ "github.com/lib/pq"
 	amqp "github.com/rabbitmq/amqp091-go"
